Move formatting logic of replacer package to its own file
